Transition to usage of index instead of scan number. Fix random access for SIM files.
The id parsing relies on the last value of the key-value pairs being a
unique number. This can't be guaranteed by the standard and causes
breakage in demultiplexing software that appends to the id. The solution
chosen is to use the index tag rather than the id tag for indexing.

A sanity check performed when finding indexList offsets fails to account
for SIM files. This caused these files to set 'seekable' to False. Fix
this.